Structure spawns and other components
This splits up the spawning components quite a bit so that it's in
multiple modules instead of 1, which means it's a bit easier to read.

In addition to that I'm shuffling some other code around and adding a
new spawnable for structures, which are _slightly_ different from the
tdlg ones, which I might want to fix up at some point in time.