Split apart jobs
This commit obviously does too much. But as a summary, it splits apart
the jobs to different components, so that it will be easier for the
assignments to target only characters with specific jobs given to them.

I'm keeping the assigning of jobs to a single system, so that it all
funnels into a single place so that characters won't ever have multiple
jobs.

I might eventually swap that around, or maybe even bring back the other
way of doing this, but for now I am going to run with this.