Switch to different structs for different tasks
Instead of having these as one enum, I am splitting them into multiple
structs so that my systems that have to do the work with the characters
can target very specific tasks that need to get done and only bring in
the other information they need.

This does make it harder to assert that there is only one task per
character, but maybe that's ok and I will just have to eat that
complexity.