Spawn a second character
All of these are fixes so that the game can handle multiple characters
being in the game at the same time.

There's still a bit more to be done so that the characters don't get
stuck doing nothing while the other characters do tasks. The problem
is that I'm assigning a job priority when the character can't actually
get a task. I need to revisit the job priority and beef it up a bit to
be a bit smarter.